+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New functions
 
 * `get_variances()`, to calculate the variance components from mixed models of class `merMod`, `glmmTMB` and `stanreg` (#52).
+* `find_algorithm()`, to get information about sampling algorithms and optimizers, and for Bayesian models also about chains and iterations (#38).
 
 ## Changes to functions